move config file to proper location
