+++ src/main/java/at/hannibal2/skyhanni/config/features/event/HoppityEggsConfig.java
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.config.features.misc;
+package at.hannibal2.skyhanni.config.features.event;
 
 import at.hannibal2.skyhanni.config.FeatureToggle;
 import at.hannibal2.skyhanni.config.core.config.Position;
